requirements.txt
# Flask Backend Dependencies
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyJWT>=2.8.0
Werkzeug>=3.0.0
